requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
django-rest-knox==4.1.0
djangorestframework==3.11.0
docutils==0.15.2
ibm-cos-sdk==2.6.0
ibm-cos-sdk-core==2.6.0
ibm-cos-sdk-s3transfer==2.6.0
idna==2.5
isort==4.3.21
jmespath==0.10.0
joblib==0.15.1
lazy-object-proxy==1.4.3
lomond==0.3.3
mccabe==0.6.1
numpy==1.18.5
pandas==0.25.3
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
scikit-learn==0.20.1
scipy==1.5.0
six==1.15.0
sqlparse==0.3.1
tabulate==0.8.7
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
urllib3==1.21.1
watson-machine-learning-client==1.0.378
watson-machine-learning-client-V4==1.0.103
wrapt==1.12.1
